# FE/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sports Center Management API",
    "version": "1.0.0",
    "description": "Backend API for Sports Center Management System – Flow 1 (User & Membership), Flow 2 (Class & Schedule), Flow 3 (Payment & Report)"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://sports-center-management-system.onrender.com/api/v1",
      "description": "Production Server (Render)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request (validation or malformed body)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Validation failed",
                "errors": [
                  {
                    "field": "email",
                    "message": "Invalid email address"
                  }
                ]
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, invalid or expired token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Unauthorized: invalid or expired token"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient role permissions",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Forbidden: insufficient permissions"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Record not found"
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Internal server error"
              }
            }
          }
        }
      },
...
      "SubscriptionCreated": {
        "description": "Subscription created with its payment",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": true,
                "message": "Subscription created successfully",
                "data": {
                  "subscription": {
                    "id": "9643ec65-bacb-4b1b-9442-239bb60bd8fa",
                    "tier": "MEMBERSHIP",
                    "startDate": "2026-09-11T14:20:14.968Z",
                    "endDate": "2026-10-11T14:20:14.968Z",
                    "status": "ACTIVE",
                    "plan": {
                      "name": "Membership Monthly",
                      "tier": "MEMBERSHIP"
                    }
                  },
                  "payment": {
                    "id": "322da21d-5040-44b8-90cc-cfc9eeff2631",
                    "amount": "300000",
                    "method": "CASH",
                    "status": "SUCCESS"
                  }
                }
              }
            }
          }
        }
      },
      "SubscriptionListOk": {
        "description": "Paginated list of subscriptions (compact example)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": true,
                "message": "Subscriptions retrieved successfully",
                "data": [
                  {
                    "id": "9643ec65-bacb-4b1b-9442-239bb60bd8fa",
                    "tier": "MEMBERSHIP",
                    "startDate": "2026-09-11T14:20:14.968Z",
                    "endDate": "2026-10-11T14:20:14.968Z",
                    "status": "ACTIVE",
                    "suspendedAt": null,
                    "remainingDays": null,
                    "plan": {
                      "name": "Membership Monthly",
                      "tier": "MEMBERSHIP"
                    },
                    "payments": [
                      {
                        "id": "322da21d-5040-44b8-90cc-cfc9eeff2631",
                        "amount": "300000",
                        "status": "SUCCESS"
                      }
                    ]
                  }
                ],
                "pagination": {
                  "page": 1,
                  "limit": 10,
                  "total": 1,
                  "totalPages": 1
                }
              }
            }
          }
        }
      },
      "SubscriptionOk": {
        "description": "Single subscription",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": true,
                "message": "Subscription retrieved successfully",
                "data": {
                  "id": "9643ec65-bacb-4b1b-9442-239bb60bd8fa",
                  "tier": "MEMBERSHIP",
                  "startDate": "2026-09-11T14:20:14.968Z",
                  "endDate": "2026-10-11T14:20:14.968Z",
                  "status": "SUSPENDED",
                  "suspendedAt": "2026-09-20T14:20:14.968Z",
                  "remainingDays": 21,
                  "plan": {
                    "name": "Membership Monthly",
                    "tier": "MEMBERSHIP"
                  },
                  "member": {
                    "user": {
                      "fullName": "John Doe",
                      "email": "member1@example.com"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/subscriptions": {
      "post": {
        "summary": "Register member to a membership plan",
        "tags": [
          "Subscriptions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "memberId",
                  "planId",
                  "paymentMethod"
                ],
                "properties": {
                  "memberId": {
                    "type": "string"
                  },
                  "planId": {
                    "type": "string"
                  },
                  "startDate": {
                    "type": "string"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "CASH",
                      "BANK_TRANSFER"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/SubscriptionCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/subscriptions/{id}/renew": {
      "post": {
        "summary": "Renew a membership subscription",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "planId",
                  "paymentMethod"
                ],
                "properties": {
                  "planId": {
                    "type": "string"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "CASH",
                      "BANK_TRANSFER"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/SubscriptionCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/subscriptions/member/{memberId}": {
      "get": {
        "summary": "Get all subscriptions for a member",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "EXPIRED",
                "CANCELLED",
                "SUSPENDED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SubscriptionListOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/subscriptions/{id}": {
      "get": {
        "summary": "Get subscription by ID",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SubscriptionOk"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/subscriptions/{id}/status": {
      "patch": {
        "summary": "Cập nhật trạng thái gói (Manager only)",
        "description": "**Khi đổi sang `CANCELLED` (Manager hủy gói):**\n- Tự động tính hoàn tiền theo tỷ lệ ngày còn lại *(Prorated)*:\n  `refundAmount = (giá_gốc / tổng_ngày) × ngày_còn_lại`\n- Nếu còn ngày sử dụng → Payment gốc chuyển sang `REFUNDED`, Member nhận Notification `PAYMENT_REFUNDED`.\n- Nếu gói đã hết hạn (0 ngày còn lại) → Không hoàn tiền, gửi `SUBSCRIPTION_CANCELLED`.\n- Toàn bộ lịch học tương lai (`BOOKED`) của member **tự động bị hủy**.\n\n**Khi đổi sang `SUSPENDED`:** Lưu số ngày còn lại để khôi phục sau.\n\n**Khi đổi từ `SUSPENDED` → `ACTIVE`:** Cộng lại số ngày còn dư vào ngày kết thúc mới.\n",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "CANCELLED",
                      "SUSPENDED"
                    ]
                  }
                }
              },
              "examples": {
                "cancel": {
                  "summary": "Hủy gói (tự động tính hoàn tiền)",
                  "value": {
                    "status": "CANCELLED"
                  }
                },
                "suspend": {
                  "summary": "Tạm dừng gói",
                  "value": {
                    "status": "SUSPENDED"
                  }
                },
                "resume": {
                  "summary": "Kích hoạt lại từ tạm dừng",
                  "value": {
                    "status": "ACTIVE"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "cancelled_with_refund": {
                    "summary": "Hủy + hoàn tiền prorated",
                    "value": {
                      "success": true,
                      "message": "Subscription status updated",
                      "data": {
                        "id": "sub-uuid",
                        "status": "CANCELLED",
                        "refundAmount": 200000,
                        "willRefund": true,
                        "daysLeft": 20
                      }
                    }
                  },
                  "cancelled_no_refund": {
                    "summary": "Hủy gói đã hết hạn",
                    "value": {
                      "success": true,
                      "message": "Subscription status updated",
                      "data": {
                        "id": "sub-uuid",
                        "status": "CANCELLED",
                        "refundAmount": 0,
                        "willRefund": false,
                        "daysLeft": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/subscriptions/{id}/cancel": {
      "patch": {
        "summary": "Member tự hủy gói của mình",
        "description": "**Chính sách hoàn tiền khi Member tự hủy gói:**\n- Gói còn **> 15 ngày** sử dụng → Hoàn **30%** giá trị gói gốc (qua nhân viên).\n- Gói còn **≤ 15 ngày** sử dụng → **Không hoàn tiền**.\n\n**Hiệu lực khi hủy:**\n- Subscription chuyển sang `CANCELLED`.\n- Toàn bộ lịch học đã đặt trong tương lai (`BOOKED`) sẽ tự động bị `CANCELLED`.\n- Payment gốc chuyển sang `REFUNDED` nếu đủ điều kiện.\n- Member nhận Notification ngay lập tức với thông tin hoàn tiền (nếu có).\n",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Subscription ID (phải là gói của chính mình)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Lý do hủy gói (không bắt buộc)",
                    "example": "Tôi không có thời gian tập luyện nữa"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hủy thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "refund_eligible": {
                    "summary": "Hoàn tiền 30% (còn > 15 ngày)",
                    "value": {
                      "success": true,
                      "message": "Hủy thành công. Hoàn 150,000đ (30%) vì còn 20 ngày.",
                      "data": {
                        "subscriptionId": "sub-uuid",
                        "status": "CANCELLED",
                        "daysLeft": 20,
                        "refundAmount": 150000,
                        "willRefund": true
                      }
                    }
                  },
                  "no_refund": {
                    "summary": "Không hoàn tiền (còn ≤ 15 ngày)",
                    "value": {
                      "success": true,
                      "message": "Hủy thành công. Không hoàn tiền vì còn ≤ 15 ngày (còn 10 ngày).",
                      "data": {
                        "subscriptionId": "sub-uuid",
                        "status": "CANCELLED",
                        "daysLeft": 10,
                        "refundAmount": 0,
                        "willRefund": false
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Gói không ở trạng thái ACTIVE",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "success": false,
                  "message": "Không thể hủy gói đang ở trạng thái CANCELLED"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}